added: readme and sample files

+++ app/Traits/FileManager.php
@@ -42,7 +42,11 @@ trait FileManager
     public function deleteFile($path)
     {
         if (Storage::exists($path)) {
-            Storage::delete($path);
+            if (!in_array($path, config('default_images'))) {
+                if (!in_array($path, config('default_videos'))) {
+                    Storage::delete($path);
+                }
+            }
         }
     }
 }
